requirements.txt
alembic==1.9.4
appdirs==1.4.4
attrs==22.2.0
autocommand==2.2.2
Babel==2.11.0
bcrypt==4.0.1
blinker==1.5
Brlapi==0.8.4
Brotli==1.0.9
btrfsutil==6.2.1
cffi==1.15.1
chardet==5.1.0
cli-helpers==2.3.0
click==8.1.3
cloud-init==23.1.1
configobj==5.1.0.dev0
contourpy==1.0.7
cryptography==39.0.2
cycler==0.11.0
Cython==0.29.33
decorator==5.1.1
distlib==0.3.6
dnspython==2.3.0
docopt==0.6.2
email-validator==1.3.1
extras==1.0.0
fastjsonschema==2.16.3
filelock==3.11.0
fixtures==4.0.1
Flask==2.2.2
Flask-BabelEx==0.9.4
Flask-Compress==1.13
Flask-Gravatar==0.5.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==4.1.5
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
fonttools==4.39.2
greenlet==2.0.2
gssapi==1.8.2
idna==3.4
importlib-metadata==5.0.0
inflect==6.0.2
itsdangerous==2.1.2
jaraco.context==4.3.0
jaraco.functools==3.6.0
jaraco.text==3.11.1
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.17.3
kiwisolver==1.4.4
ldap3==2.9
lensfun==0.3.3
libfdt==1.6.1
louis==3.25.0
Mako==1.2.4
MarkupSafe==2.1.2
matplotlib==3.6.3
more-itertools==9.1.0
mycli==1.25.0
mysql-connector-python==8.0.33
netifaces==0.11.0
nftables==0.1
nspektr==0.4.0
numpy==1.24.2
oauthlib==3.2.2
ordered-set==4.1.0
packaging==23.0
paramiko==2.11.1
passlib==1.7.4
pbr==5.11.1
Pillow==9.4.0
platformdirs==3.1.0
ply==3.11
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
psycopg2==2.9.5
pwquality==1.4.5
pyaes==1.6.1
pyasn1==0.4.8
pybind11==2.10.4
pycairo==1.23.0
pycparser==2.21
pydantic==1.10.6
Pygments==2.14.0
PyGObject==3.42.2
PyMySQL==1.0.2
PyNaCl==1.4.0
pyparsing==3.0.9
pyperclip==1.8.2
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyrsistent==0.19.3
pyserial==3.5
python-dateutil==2.8.2
python-editor==1.0.4
python-xlib==0.33
pytz==2022.7.1
pywal==3.3.0
PyYAML==6.0
ranger-fm==1.9.3
Reflector==2021.11.20.2.41.3
requests==2.28.2
rfc3987==1.3.8
simplejson==3.18.0
six==1.16.0
speaklater==1.3
SQLAlchemy==1.3.24
sqlparse==0.4.3
sshtunnel==0.4.0
tabulate==0.9.0
team==1.0
testtools==2.5.0
tomli==2.0.1
trash-cli==0.23.2.13.2
trove-classifiers==2023.3.10
typing_extensions==4.5.0
ueberzug==18.1.9
urllib3==1.26.13
validate==5.1.0.dev0
validate-pyproject==0.12.1
virtualenv==20.21.0
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
